requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.3
django-cors-headers==4.0.0
djangorestframework==3.14.0
idna==3.4
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.7.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
telebot-calendar==1.2
tzdata==2022.7
urllib3==1.26.15

